requirements.txt
django
django-rest-framework
requests
beautifulsoup4
psycopg2
pytest
pytest-django
pytest-mock
python-dotenv